Extra warning about use of SSH add-on
I tried to make it clearer that the SSH setup here (port 22222) was different to the SSH add-on setup (port 22) and what issues would crop up if the later was used in this context. 

It took me a while to figure out what was going on and I felt that the doc needed this clarification.