Update Docker requirements
With the Windows 10 20H1 going to the Release Preview, WSL 2 is now available in a stable form. WSL 2 is supported by edge builds of Docker Desktop and so we can use devcontainers on WIndows 10 Home 🎉

Little tweak to the instructions to cover this case. When 20H1 goes to release (May?) WSL 2 should probably be the recommended method for all levels of Windows.

Also, a minor tweak to the heading levels as Docusaurus assumes `#` are Page headings and doesn't put them in the TOC on the right. 

Also also, a minor correction to capitalisation of headings for consistency