Adding TODO for original author(s) to update this documentation
TODO: The instructions below are extremely unclear and it is unknown what steps to take. See recent conversations on this outlining that every step of these instructions are non-sensical, out of date, and/or do not reflect any steps or methods one would use to follow them. See: https://community.home-assistant.io/t/debugging-the-home-assistant-operating-system/197829

"You need set the dev mode on supervisor and enable debug with options. You need also install the Remote debug Add-on from Developer Repository to expose the debug port to Host." ????