Added semantic tags to the introductory page
I have added the paragraph tags along with heading tags to improve the semantic meaning of the page. In addition to this, the unordered list is also created so that it becomes easy for the screen readers to process the data. Semantic tags of bold and italic text have also been added.