Added Python Example
Added a Python POST example underneath the `/api/services/<domain>/<service>` section.
I have been learning Python and I noticed that there was no Python example in that section, a few months ago I was looking myself for a Python example in that section. I've tried to close follow other examples in their layout and format. The code is tested and working.