Describe how to enable debugging in VS Code within Dev container
Add a step-by-step guideline to enable debugging in VS Code and Dev container and how to ensure it's set up correctly. This is something that is missing in the dev docs and for the newbies it's not obvious as debugging doesn't work out-of-the-box after setting up the dev container in VS Code.