Add notes about debugging HA in Dev Container
Since some people (like me and others) experience issues with the default Dev Container debugging, whcih doesn't work out-of-the box - new developers may think that debugging is not supported by the dev container out-of-the-box. This description is to indicate that normally debugging doesn't needs any additional actions to be taken to enable it. I also mentioned the way to debug remote instances, which may be very helpful for the developers, also in case of issues with the default debugging without enabling `debugpy`.
If I had such description I wouldn't loose almost 2 hours of googling trying to find out what is supported and what are the options, so I hope thanks to this people will avoid this.