Mention mirrored networking mode for WSL (#2586)
* Mention mirrored networking mode for WSL

Mirrorred networking mode is great for Home Assistant because it allows (among other things) working mDNS device discovery.

I think it is incredibly useful so wanted to highlight it. I think that means putting it in a "tip", but there are now 2 tips which is not great. And technically the existing Tip is not needed anymore when Mirrored networking mode is used. 

Not sure how to better organize it. Suggestions are welcome.

* Remove leftover link and tweak text a bit