Do not use sudo for running OS's scripts/enter.sh
Do not suggest to use `sudo` when running OS build - the container is
started as privileged one and all the operations on host which needs
sudo right are invoked with explicit `sudo`. If executed with sudo,
configure now refuses to run because of a sanity check:

configure: error: you should not run configure as root (set FORCE_UNSAFE_CONFIGURE=1 in environment to bypass this check)

Also update output samples to mimic current Docker's output.

Refs home-assistant/operating-system#4214