Migrate HAOS docs from the OS repo
Move documentation from the OS repository to the developer docs. All of
the docs tracked in the repo were generally developer-facing and some
information was documented here, some in the repo.

As the first step, move all the docs as-is with only the links modified
and structure adjusted for the rendered docs, without further
proofreading or checks (TBD later).