Add-ons: Bind mount Docker config only
The goal of bind mounting the ~/.docker directory is to have the
credentials available inside the builder. However, when using buildx
some files get altered in the home directory with root privileges, which
can be problematic when using Docker commands without sudo outside of
the context of the builder.

Instead of bind mounting the complete ~/.docker directory simply bind
mount config.json only which has the essential credentials.