Fix broken links in voice docs
Depending on how you navigate to a particular page the URL may end with a
`/` or not. For example, on the 3 pages updated here, navigating to them
by clicking a link in the side nav results in a URL without the trailing
slash, but refreshing the page or visiting the URL directly (e.g. from a
bookmark) results in a URL with a trailing slash.

This causes problems with relative URLs as they assume a fixed number of
path segments in the URL.

Update relative URLs to start from the root (i.e. `/docs`) to ensure
consistency regardless of whether the URLs end with a slash or not.
Some of the other links in the voice docs already use this approach.