Fix example on how-to delete unused options
The bashio config helper functions use the Supervisor add-on config API
which seems to filter options according to the current schema. This
means that after removal of an option, the config will no longer be
presented, despite it being still present in the Supervisors option
dictionary.

Instead, use the Supervisor option API to read all currently stored
options. This returns all configs, even those which are no longer
present.